LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v2/script-src/manifest-v2-script-src-blob-scheme-allowed-expected.txt

ALERT: PASS executed blob-URL script.
ALERT: PASS successfully loaded blob-URL script.

Tests CSP for extensions permits the blob: scheme.
